split BasicCegarLoop into abstract class and concrete NwaCegarLoop
