+++ trunk/source/TraceAbstraction/src/de/uni_freiburg/informatik/ultimate/plugins/generator/traceabstraction/BasicCegarLoop.java
@@ -32,18 +32,11 @@ package de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import de.uni_freiburg.informatik.ultimate.automata.AutomataLibraryException;
 import de.uni_freiburg.informatik.ultimate.automata.AutomataLibraryServices;
@@ -53,7 +46,6 @@ import de.uni_freiburg.informatik.ultimate.automata.IAutomaton;
 import de.uni_freiburg.informatik.ultimate.automata.IRun;
 import de.uni_freiburg.informatik.ultimate.automata.Word;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.EpsilonNestedWordAutomaton;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.IDoubleDeckerAutomaton;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.INestedWordAutomaton;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.INwaOutgoingLetterAndTransitionProvider;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.NestedRun;
@@ -61,53 +53,34 @@ import de.uni_freiburg.informatik.ultimate.automata.nestedword.NestedWord;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.NestedWordAutomaton;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.VpAlphabet;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.Accepts;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.Difference;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.Intersect;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.IsEmpty;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.IsEmpty.SearchStrategy;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.IsEmptyHeuristic;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.IsEmptyHeuristic.AStarHeuristic;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.IsEmptyHeuristic.IHeuristic;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.PowersetDeterminizer;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.RemoveDeadEnds;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.RemoveUnreachable;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.operations.oldapi.IOpWithDelayedDeadEndRemoval;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.senwa.DifferenceSenwa;
 import de.uni_freiburg.informatik.ultimate.core.lib.exceptions.RunningTaskInfo;
 import de.uni_freiburg.informatik.ultimate.core.lib.exceptions.TaskCanceledException;
 import de.uni_freiburg.informatik.ultimate.core.lib.exceptions.TaskCanceledException.UserDefinedLimit;
 import de.uni_freiburg.informatik.ultimate.core.lib.exceptions.ToolchainCanceledException;
-import de.uni_freiburg.informatik.ultimate.core.lib.results.DangerInvariantResult;
 import de.uni_freiburg.informatik.ultimate.core.model.models.IElement;
 import de.uni_freiburg.informatik.ultimate.core.model.preferences.IPreferenceProvider;
 import de.uni_freiburg.informatik.ultimate.core.model.services.IUltimateServiceProvider;
 import de.uni_freiburg.informatik.ultimate.core.model.translation.IProgramExecution;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.CfgSmtToolkit;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.IcfgProgramExecution;
-import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.IcfgUtils;
-import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.structure.IAction;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.structure.IIcfg;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.structure.IIcfgTransition;
-import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.structure.IcfgEdge;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.structure.IcfgLocation;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.structure.debugidentifiers.DebugIdentifier;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.hoaretriple.HoareTripleCheckerUtils;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.hoaretriple.IHoareTripleChecker;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.hoaretriple.IncrementalHoareTripleChecker;
-import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.smt.predicates.IMLPredicate;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.smt.predicates.IPredicate;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.smt.predicates.IPredicateUnifier;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.smt.predicates.ISLPredicate;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.smt.predicates.PredicateFactory;
-import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.smt.predicates.PredicateUnifier;
-import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.smt.tracecheck.ITraceCheckPreferences.AssertCodeBlockOrderType;
-import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.smt.tracecheck.ITraceCheckPreferences.UnsatCores;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.taskidentifier.SubtaskIterationIdentifier;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.tracehandling.IRefinementEngineResult;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.tracehandling.IRefinementEngineResult.BasicRefinementEngineResult;
-import de.uni_freiburg.informatik.ultimate.lib.smtlibutils.SmtUtils.SimplificationTechnique;
-import de.uni_freiburg.informatik.ultimate.lib.smtlibutils.SmtUtils.XnfConversionTechnique;
-import de.uni_freiburg.informatik.ultimate.lib.smtlibutils.solverbuilder.SMTFeatureExtractionTermClassifier.ScoringMethod;
 import de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.cfg2automaton.Cfg2Automaton;
 import de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.singletracecheck.InterpolationTechnique;
 import de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.singletracecheck.TraceCheckUtils;
@@ -116,24 +89,16 @@ import de.uni_freiburg.informatik.ultimate.logic.Term;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.rcfgbuilder.cfg.PathProgram;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.rcfgbuilder.cfg.PathProgram.PathProgramConstructionResult;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.rcfgbuilder.util.IcfgAngelicProgramExecution;
-import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.automataminimization.AutomataMinimization;
-import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.automataminimization.AutomataMinimization.AutomataMinimizationTimeout;
-import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.errorabstraction.ErrorGeneralizationEngine;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.errorlocalization.FlowSensitiveFaultLocalizer;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.interpolantautomata.transitionappender.AbstractInterpolantAutomaton;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.interpolantautomata.transitionappender.DeterministicInterpolantAutomaton;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.interpolantautomata.transitionappender.NondeterministicInterpolantAutomaton;
-import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.pathinvariants.PathInvariantsGenerator;
-import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.pathinvariants.internal.DangerInvariantGuesser;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.predicates.InductivityCheck;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.preferences.TAPreferences;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.preferences.TAPreferences.Artifact;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.preferences.TAPreferences.InterpolantAutomatonEnhancement;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.preferences.TraceAbstractionPreferenceInitializer;
-import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.preferences.TraceAbstractionPreferenceInitializer.CounterexampleSearchStrategy;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.preferences.TraceAbstractionPreferenceInitializer.FloydHoareAutomataReuse;
-import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.preferences.TraceAbstractionPreferenceInitializer.InterpolantAutomaton;
-import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.preferences.TraceAbstractionPreferenceInitializer.Minimization;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.preferences.TraceAbstractionPreferenceInitializer.RelevanceAnalysisMode;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.tracehandling.StrategyFactory;
 import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.tracehandling.TaCheckAndRefinementPreferences;
@@ -151,8 +116,11 @@ import de.uni_freiburg.informatik.ultimate.util.statistics.IStatisticsDataProvid
  *
  * @author Matthias Heizmann (heizmann@informatik.uni-freiburg.de)
  * @author Christian Schilling (schillic@informatik.uni-freiburg.de)
+ *
+ * @param <A>
+ *            The type of abstraction refined by the CEGAR loop
  */
-public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L, IPredicate>>
+public abstract class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L, IPredicate>>
 		extends AbstractCegarLoop<L, A> {
 
 	public enum AutomatonType {
@@ -182,34 +150,16 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 
 	}
 
-	protected static final int MINIMIZE_EVERY_KTH_ITERATION = 10;
-	protected static final boolean REMOVE_DEAD_ENDS = true;
-	protected static final int MINIMIZATION_TIMEOUT = 1_000;
 	private static final boolean NON_EA_INDUCTIVITY_CHECK = false;
-
-	/**
-	 * If the trace histogram max is larger than this number we try to find a danger invariant. Only used for debugging.
-	 */
-	private static final int DEBUG_DANGER_INVARIANTS_THRESHOLD = Integer.MAX_VALUE;
 
 	protected final PredicateFactoryRefinement mStateFactoryForRefinement;
 	protected final PredicateFactoryForInterpolantAutomata mPredicateFactoryInterpolantAutomata;
 	protected final PredicateFactoryResultChecking mPredicateFactoryResultChecking;
 
-	protected final InterpolantAutomaton mInterpolantAutomatonConstructionProcedure;
-	protected final UnsatCores mUnsatCores;
-	protected final boolean mUseLiveVariables;
-
-	protected final AssertCodeBlockOrderType mAssertCodeBlocksIncrementally;
-	protected final Collection<INwaOutgoingLetterAndTransitionProvider<L, IPredicate>> mStoredRawInterpolantAutomata;
-
-	private final RelevanceAnalysisMode mFaultLocalizationMode;
+	protected final RelevanceAnalysisMode mFaultLocalizationMode;
 	private final boolean mFaultLocalizationAngelic;
-	private final Set<IcfgLocation> mHoareAnnotationLocations;
-	private final SearchStrategy mSearchStrategy;
 	private final StrategyFactory<L> mStrategyFactory;
 	private final PathProgramDumpController<L> mPathProgramDumpController;
-	private final ErrorGeneralizationEngine<L> mErrorGeneralizationEngine;
 	private final boolean mStoreFloydHoareAutomata;
 	private final Set<Pair<AbstractInterpolantAutomaton<L>, IPredicateUnifier>> mFloydHoareAutomata =
 			new LinkedHashSet<>();
@@ -219,10 +169,6 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 	protected IRefinementEngineResult<L, NestedWordAutomaton<L, IPredicate>> mRefinementResult;
 
 	private boolean mFirstReuseDump = true;
-	private boolean mUseHeuristicEmptinessCheck = false;
-	private final ScoringMethod mScoringMethod;
-	private final AStarHeuristic mAStarHeuristic;
-	private final Integer mAStarRandomHeuristicSeed;
 
 	public BasicCegarLoop(final DebugIdentifier name, final A initialAbstraction, final IIcfg<?> rootNode,
 			final CfgSmtToolkit csToolkit, final PredicateFactory predicateFactory, final TAPreferences taPrefs,
@@ -239,19 +185,12 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 			interpolation = InterpolationTechnique.ForwardPredicates;
 		}
 
-		mInterpolantAutomatonConstructionProcedure = mPref.interpolantAutomaton();
-		mHoareAnnotationLocations = hoareAnnotationLocs;
 		mStoreFloydHoareAutomata = taPrefs.getFloydHoareAutomataReuse() != FloydHoareAutomataReuse.NONE;
-		mErrorGeneralizationEngine = new ErrorGeneralizationEngine<>(services);
-		mHaf = new HoareAnnotationFragments<>(mLogger, mHoareAnnotationLocations, mPref.getHoareAnnotationPositions());
+		mHaf = new HoareAnnotationFragments<>(mLogger, hoareAnnotationLocs, mPref.getHoareAnnotationPositions());
 		mStateFactoryForRefinement = stateFactoryForRefinement;
 
 		mPredicateFactoryInterpolantAutomata = new PredicateFactoryForInterpolantAutomata(
 				super.mCsToolkit.getManagedScript(), mPredicateFactory, computeHoareAnnotation);
-
-		mAssertCodeBlocksIncrementally = getServices().getPreferenceProvider(Activator.PLUGIN_ID).getEnum(
-				TraceAbstractionPreferenceInitializer.LABEL_ASSERT_CODEBLOCKS_INCREMENTALLY,
-				AssertCodeBlockOrderType.class);
 
 		mPredicateFactoryResultChecking = new PredicateFactoryResultChecking(mPredicateFactory);
 
@@ -259,16 +198,11 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 		mCegarLoopBenchmark.start(CegarLoopStatisticsDefinitions.OverallTime.toString());
 
 		final IPreferenceProvider prefs = getServices().getPreferenceProvider(Activator.PLUGIN_ID);
-		mUnsatCores = prefs.getEnum(TraceAbstractionPreferenceInitializer.LABEL_UNSAT_CORES, UnsatCores.class);
-		mUseLiveVariables = prefs.getBoolean(TraceAbstractionPreferenceInitializer.LABEL_LIVE_VARIABLES);
 		mFaultLocalizationMode =
 				prefs.getEnum(TraceAbstractionPreferenceInitializer.LABEL_ERROR_TRACE_RELEVANCE_ANALYSIS_MODE,
 						RelevanceAnalysisMode.class);
 		mFaultLocalizationAngelic =
 				prefs.getBoolean(TraceAbstractionPreferenceInitializer.LABEL_ERROR_TRACE_ANGELIC_VERIFICATION_ACTIVE);
-
-		mSearchStrategy = getSearchStrategy(prefs);
-		mStoredRawInterpolantAutomata = checkStoreCounterExamples(mPref) ? new ArrayList<>() : null;
 
 		final TaCheckAndRefinementPreferences<L> taCheckAndRefinementPrefs =
 				new TaCheckAndRefinementPreferences<>(getServices(), mPref, interpolation, mSimplificationTechnique,
@@ -293,130 +227,6 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 				}
 			}
 		}
-		// Heuristic Emptiness Check
-		mUseHeuristicEmptinessCheck = taPrefs.useHeuristicEmptinessCheck();
-		mScoringMethod = taPrefs.getHeuristicEmptinessCheckScoringMethod();
-		mAStarHeuristic = taPrefs.getHeuristicEmptinessCheckAStarHeuristic();
-		mAStarRandomHeuristicSeed = taPrefs.getHeuristicEmptinessCheckAStarHeuristicRandomSeed();
-	}
-
-	@Override
-	protected boolean isAbstractionEmpty() throws AutomataOperationCanceledException {
-		final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> abstraction =
-				(INwaOutgoingLetterAndTransitionProvider<L, IPredicate>) mAbstraction;
-
-		mCegarLoopBenchmark.start(CegarLoopStatisticsDefinitions.EmptinessCheckTime);
-		try {
-			if (mUseHeuristicEmptinessCheck) {
-				mCounterexample = new IsEmptyHeuristic<>(new AutomataLibraryServices(getServices()), abstraction,
-						IHeuristic.getHeuristic(mAStarHeuristic, mScoringMethod, mAStarRandomHeuristicSeed))
-								.getNestedRun();
-
-				assert checkIsEmptyHeuristic(abstraction) : "IsEmptyHeuristic did not match IsEmpty";
-			} else {
-				mCounterexample =
-						new IsEmpty<>(new AutomataLibraryServices(getServices()), abstraction, mSearchStrategy)
-								.getNestedRun();
-			}
-		} finally {
-			mCegarLoopBenchmark.stop(CegarLoopStatisticsDefinitions.EmptinessCheckTime);
-		}
-
-		if (mCounterexample == null) {
-			return true;
-		}
-
-		if (mPref.dumpAutomata()) {
-			mCegarLoopBenchmark.start(CegarLoopStatisticsDefinitions.DumpTime);
-			mDumper.dumpNestedRun(mCounterexample);
-			mCegarLoopBenchmark.stop(CegarLoopStatisticsDefinitions.DumpTime);
-		}
-		mLogger.info("Found error trace");
-
-		if (mLogger.isDebugEnabled()) {
-			mLogger.debug(mCounterexample.getWord());
-		}
-		final HistogramOfIterable<L> traceHistogram = new HistogramOfIterable<>(mCounterexample.getWord());
-		mCegarLoopBenchmark.reportTraceHistogramMaximum(traceHistogram.getMax());
-		if (mLogger.isInfoEnabled()) {
-			mLogger.info("trace histogram " + traceHistogram.toString());
-		}
-		if (traceHistogram.getMax() > DEBUG_DANGER_INVARIANTS_THRESHOLD) {
-			checkForDangerInvariantAndReport();
-		}
-
-		if (mPref.hasLimitTraceHistogram() && traceHistogram.getMax() > mPref.getLimitTraceHistogram()) {
-			final String taskDescription =
-					"bailout by trace histogram " + traceHistogram.toString() + " in iteration " + mIteration;
-			throw new TaskCanceledException(UserDefinedLimit.TRACE_HISTOGRAM, getClass(), taskDescription);
-		}
-
-		return false;
-	}
-
-	private boolean checkIsEmptyHeuristic(final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> abstraction)
-			throws AutomataOperationCanceledException {
-		final NestedRun<L, IPredicate> isEmptyHeuristicCex = (NestedRun<L, IPredicate>) mCounterexample;
-		final NestedRun<L, IPredicate> isEmptyCex =
-				new IsEmpty<>(new AutomataLibraryServices(getServices()), abstraction, mSearchStrategy).getNestedRun();
-
-		final Function<NestedRun<L, IPredicate>, String> toStr =
-				a -> a.getWord().asList().stream().map(b -> "T" + b.hashCode()).collect(Collectors.joining(" "));
-
-		if (isEmptyHeuristicCex == null && isEmptyCex == null) {
-			return true;
-		}
-		if (isEmptyHeuristicCex != null && isEmptyCex == null) {
-			mLogger.fatal("IsEmptyHeuristic found a path but IsEmpty did not.");
-			mLogger.fatal("IsEmptyHeuristic: " + toStr.apply(isEmptyHeuristicCex));
-			return false;
-		}
-		if (isEmptyHeuristicCex == null && isEmptyCex != null) {
-			mLogger.fatal("IsEmptyHeuristic found no path but IsEmpty did.");
-			mLogger.fatal("IsEmpty         : " + toStr.apply(isEmptyCex));
-			return false;
-		}
-		if (isEmptyHeuristicCex != null && isEmptyCex != null) {
-			if (!NestedRun.isEqual(isEmptyHeuristicCex, isEmptyCex)) {
-				if (isEmptyHeuristicCex.getLength() > isEmptyCex.getLength()) {
-					mLogger.warn("IsEmptyHeuristic and IsEmpty found a path, but isEmptyHeuristic was longer!");
-				} else {
-					mLogger.info("IsEmptyHeuristic and IsEmpty found a path, but they differ");
-				}
-				mLogger.info("IsEmptyHeuristic: " + toStr.apply(isEmptyHeuristicCex));
-				mLogger.info("IsEmpty         : " + toStr.apply(isEmptyCex));
-			}
-			return true;
-		}
-		mLogger.fatal("Should not happen");
-		return false;
-	}
-
-	private boolean checkForDangerInvariantAndReport() {
-		final Set<? extends IcfgEdge> allowedTransitions = PathInvariantsGenerator.extractTransitionsFromRun(
-				(NestedRun<? extends IAction, IPredicate>) mCounterexample,
-				mIcfg.getCfgSmtToolkit().getIcfgEdgeFactory());
-		final PathProgramConstructionResult ppResult =
-				PathProgram.constructPathProgram("PathInvariantsPathProgram", mIcfg, allowedTransitions);
-		final IIcfg<IcfgLocation> pathProgram = ppResult.getPathProgram();
-		final PredicateFactory predicateFactory = mPredicateFactory;
-		final IPredicateUnifier predicateUnifier = new PredicateUnifier(mLogger, getServices(),
-				mCsToolkit.getManagedScript(), predicateFactory, mCsToolkit.getSymbolTable(),
-				SimplificationTechnique.SIMPLIFY_DDA, XnfConversionTechnique.BOTTOM_UP_WITH_LOCAL_SIMPLIFICATION);
-		final IPredicate precondition = predicateUnifier.getTruePredicate();
-		final DangerInvariantGuesser dig = new DangerInvariantGuesser(pathProgram, getServices(), precondition,
-				predicateFactory, predicateUnifier, mCsToolkit);
-		final boolean hasDangerInvariant = dig.isDangerInvariant();
-		if (hasDangerInvariant) {
-			final Map<IcfgLocation, IPredicate> invarP = dig.getCandidateInvariant();
-			final Map<IcfgLocation, Term> invarT =
-					invarP.entrySet().stream().collect(Collectors.toMap(Entry::getKey, x -> x.getValue().getFormula()));
-			final Set<IcfgLocation> errorLocations = IcfgUtils.getErrorLocations(pathProgram);
-			final DangerInvariantResult<?, Term> res = new DangerInvariantResult<>(Activator.PLUGIN_ID, invarT,
-					errorLocations, getServices().getBacktranslationService());
-			getServices().getResultService().reportResult(Activator.PLUGIN_ID, res);
-		}
-		return hasDangerInvariant;
 	}
 
 	@Override
@@ -526,7 +336,7 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 				new IncrementalHoareTripleChecker(super.mCsToolkit, false)).getResult();
 	}
 
-	private static boolean
+	protected static boolean
 			isInterpolantAutomatonOfSingleStateType(final INestedWordAutomaton<?, IPredicate> automaton) {
 		Class<? extends IPredicate> typeofState = null;
 		for (final IPredicate state : automaton.getStates()) {
@@ -541,23 +351,7 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 	}
 
 	@Override
-	protected void constructErrorAutomaton() throws AutomataOperationCanceledException {
-		mErrorGeneralizationEngine.constructErrorAutomaton(mCounterexample, mPredicateFactory,
-				mRefinementResult.getPredicateUnifier(), mCsToolkit, mSimplificationTechnique, mXnfConversionTechnique,
-				mIcfg.getCfgSmtToolkit().getSymbolTable(), mPredicateFactoryInterpolantAutomata,
-				(INestedWordAutomaton<L, IPredicate>) mAbstraction, mIteration);
-
-		mInterpolAutomaton = null;
-		final NestedWordAutomaton<L, IPredicate> resultBeforeEnhancement =
-				mErrorGeneralizationEngine.getResultBeforeEnhancement();
-		assert isInterpolantAutomatonOfSingleStateType(resultBeforeEnhancement);
-		assert accepts(getServices(), resultBeforeEnhancement, mCounterexample.getWord(),
-				false) : "Error automaton broken!";
-	}
-
-	@Override
 	protected void finish() {
-		mErrorGeneralizationEngine.reportErrorGeneralizationBenchmarks();
 		final List<Integer> sortedHistogram = mStrategyFactory.getPathProgramCache().computeSortedHistrogram();
 		mLogger.info("Path program histogram: " + sortedHistogram);
 		final int max = HistogramOfIterable.getMaxOfVisualizationArray(sortedHistogram);
@@ -573,52 +367,6 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 		}
 		return HoareTripleCheckerUtils.constructEfficientHoareTripleCheckerWithCaching(getServices(),
 				mPref.getHoareTripleChecks(), mCsToolkit, mRefinementResult.getPredicateUnifier());
-	}
-
-	@Override
-	protected boolean refineAbstraction() throws AutomataLibraryException {
-		mStateFactoryForRefinement.setIteration(mIteration);
-		mCegarLoopBenchmark.start(CegarLoopStatisticsDefinitions.AutomataDifference.toString());
-
-		final INestedWordAutomaton<L, IPredicate> minuend = (INestedWordAutomaton<L, IPredicate>) mAbstraction;
-
-		final IPredicateUnifier predicateUnifier = mRefinementResult.getPredicateUnifier();
-		final IHoareTripleChecker htc = getHoareTripleChecker();
-
-		final AutomatonType automatonType;
-		final boolean useErrorAutomaton;
-		final NestedWordAutomaton<L, IPredicate> subtrahendBeforeEnhancement;
-		final InterpolantAutomatonEnhancement enhanceMode;
-		final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> subtrahend;
-		final boolean exploitSigmaStarConcatOfIa;
-		if (mErrorGeneralizationEngine.hasAutomatonInIteration(mIteration)) {
-			mErrorGeneralizationEngine.startDifference();
-			automatonType = AutomatonType.ERROR;
-			useErrorAutomaton = true;
-			exploitSigmaStarConcatOfIa = false;
-			enhanceMode = mErrorGeneralizationEngine.getEnhancementMode();
-			subtrahendBeforeEnhancement = mErrorGeneralizationEngine.getResultBeforeEnhancement();
-			subtrahend = mErrorGeneralizationEngine.getResultAfterEnhancement();
-		} else {
-			automatonType = AutomatonType.FLOYD_HOARE;
-			useErrorAutomaton = false;
-			exploitSigmaStarConcatOfIa = !mComputeHoareAnnotation;
-			subtrahendBeforeEnhancement = mInterpolAutomaton;
-			enhanceMode = mPref.interpolantAutomatonEnhancement();
-			subtrahend = enhanceInterpolantAutomaton(enhanceMode, predicateUnifier, htc, subtrahendBeforeEnhancement);
-		}
-
-		// TODO: HTC and predicateunifier statistics are saved in the following method, but it seems better to save them
-		// at the end of the htc lifecycle instead of there
-		computeAutomataDifference(minuend, subtrahend, subtrahendBeforeEnhancement, predicateUnifier,
-				exploitSigmaStarConcatOfIa, htc, enhanceMode, useErrorAutomaton, automatonType);
-
-		minimizeAbstractionIfEnabled();
-
-		final boolean stillAccepted = new Accepts<>(new AutomataLibraryServices(getServices()),
-				(INwaOutgoingLetterAndTransitionProvider<L, IPredicate>) mAbstraction,
-				(NestedWord<L>) mCounterexample.getWord()).getResult();
-		return !stillAccepted;
 	}
 
 	protected INwaOutgoingLetterAndTransitionProvider<L, IPredicate> enhanceInterpolantAutomaton(
@@ -676,114 +424,6 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 		return new IsEmpty<>(new AutomataLibraryServices(getServices()), intersection).getResult();
 	}
 
-	private void computeAutomataDifference(final INestedWordAutomaton<L, IPredicate> minuend,
-			final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> subtrahend,
-			final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> subtrahendBeforeEnhancement,
-			final IPredicateUnifier predicateUnifier, final boolean explointSigmaStarConcatOfIA,
-			final IHoareTripleChecker htc, final InterpolantAutomatonEnhancement enhanceMode,
-			final boolean useErrorAutomaton, final AutomatonType automatonType)
-			throws AutomataLibraryException, AssertionError {
-		try {
-			mLogger.debug("Start constructing difference");
-			final PowersetDeterminizer<L, IPredicate> psd =
-					new PowersetDeterminizer<>(subtrahend, true, mPredicateFactoryInterpolantAutomata);
-			IOpWithDelayedDeadEndRemoval<L, IPredicate> diff;
-			try {
-				if (mPref.differenceSenwa()) {
-					diff = new DifferenceSenwa<>(new AutomataLibraryServices(getServices()), mStateFactoryForRefinement,
-							minuend, subtrahend, psd, false);
-				} else {
-					diff = new Difference<>(new AutomataLibraryServices(getServices()), mStateFactoryForRefinement,
-							minuend, subtrahend, psd, explointSigmaStarConcatOfIA);
-				}
-				mCegarLoopBenchmark.reportInterpolantAutomatonStates(subtrahend.size());
-			} catch (final AutomataOperationCanceledException | ToolchainCanceledException tce) {
-				final RunningTaskInfo runningTaskInfo = executeDifferenceTimeoutActions(minuend, subtrahend,
-						subtrahendBeforeEnhancement, automatonType);
-				tce.addRunningTaskInfo(runningTaskInfo);
-				throw tce;
-			} finally {
-				if (enhanceMode != InterpolantAutomatonEnhancement.NONE) {
-					assert subtrahend instanceof AbstractInterpolantAutomaton : "if enhancement is used, we need AbstractInterpolantAutomaton";
-					((AbstractInterpolantAutomaton<L>) subtrahend).switchToReadonlyMode();
-				}
-			}
-
-			if (mErrorGeneralizationEngine.hasAutomatonInIteration(mIteration)) {
-				mErrorGeneralizationEngine.stopDifference(minuend, mPredicateFactoryInterpolantAutomata,
-						mPredicateFactoryResultChecking, mCounterexample, false);
-				if (mFaultLocalizationMode != RelevanceAnalysisMode.NONE) {
-					final INestedWordAutomaton<L, IPredicate> cfg = Cfg2Automaton.constructAutomatonWithSPredicates(
-							getServices(), super.mIcfg, mStateFactoryForRefinement, super.mErrorLocs,
-							mPref.interprocedural(), mPredicateFactory);
-					mErrorGeneralizationEngine.faultLocalizationWithStorage(cfg, mCsToolkit, mPredicateFactory,
-							mRefinementResult.getPredicateUnifier(), mSimplificationTechnique, mXnfConversionTechnique,
-							mIcfg.getCfgSmtToolkit().getSymbolTable(), null, (NestedRun<L, IPredicate>) mCounterexample,
-							(IIcfg<IcfgLocation>) mIcfg);
-				}
-			}
-
-			if (mPref.dumpAutomata()) {
-				final String filename =
-						new SubtaskIterationIdentifier(mTaskIdentifier, getIteration()) + "AbstractionAfterDifference";
-				super.writeAutomatonToFile(subtrahend, filename);
-			}
-			dumpOrAppendAutomatonForReuseIfEnabled(subtrahend, predicateUnifier);
-
-			if (!useErrorAutomaton) {
-				checkEnhancement(subtrahendBeforeEnhancement, subtrahend);
-			}
-
-			if (REMOVE_DEAD_ENDS) {
-				if (mComputeHoareAnnotation) {
-					final Difference<L, IPredicate> difference = (Difference<L, IPredicate>) diff;
-					mHaf.updateOnIntersection(difference.getFst2snd2res(), difference.getResult());
-				}
-				diff.removeDeadEnds();
-				if (mComputeHoareAnnotation) {
-					mHaf.addDeadEndDoubleDeckers(diff);
-				}
-			}
-			mAbstraction = (A) diff.getResult();
-			if (mPref.dumpAutomata()) {
-				final String filename = new SubtaskIterationIdentifier(mTaskIdentifier, getIteration())
-						+ "AbstractionAfterDifferenceAndDeadEndRemoval";
-				super.writeAutomatonToFile(mAbstraction, filename);
-			}
-
-		} finally {
-			mLogger.info(predicateUnifier.collectPredicateUnifierStatistics());
-			mLogger.info(htc.getStatistics());
-			mLogger.info(htc);
-			mCegarLoopBenchmark.addEdgeCheckerData(htc.getStatistics());
-			mCegarLoopBenchmark.addPredicateUnifierData(predicateUnifier.getPredicateUnifierBenchmark());
-			mCegarLoopBenchmark.stop(CegarLoopStatisticsDefinitions.AutomataDifference.toString());
-		}
-	}
-
-	private RunningTaskInfo executeDifferenceTimeoutActions(final INestedWordAutomaton<L, IPredicate> minuend,
-			final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> subtrahend,
-			final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> subtrahendBeforeEnhancement,
-			final AutomatonType automatonType) throws AutomataLibraryException {
-		final RunningTaskInfo runningTaskInfo =
-				getDifferenceTimeoutRunningTaskInfo(minuend, subtrahend, subtrahendBeforeEnhancement, automatonType);
-		if (mErrorGeneralizationEngine.hasAutomatonInIteration(mIteration)) {
-			mErrorGeneralizationEngine.stopDifference(minuend, mPredicateFactoryInterpolantAutomata,
-					mPredicateFactoryResultChecking, mCounterexample, true);
-		}
-		return runningTaskInfo;
-	}
-
-	private RunningTaskInfo getDifferenceTimeoutRunningTaskInfo(final INestedWordAutomaton<L, IPredicate> minuend,
-			final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> subtrahend,
-			final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> subtrahendBeforeEnhancement,
-			final AutomatonType automatonType) {
-		final String taskDescription = "constructing difference of abstraction (" + minuend.size() + "states) and "
-				+ automatonType + " automaton (currently " + subtrahend.size() + " states, "
-				+ subtrahendBeforeEnhancement.size() + " states before enhancement)";
-		return new RunningTaskInfo(getClass(), taskDescription);
-	}
-
 	protected void dumpOrAppendAutomatonForReuseIfEnabled(
 			final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> automaton,
 			final IPredicateUnifier predicateUnifier) {
@@ -823,7 +463,7 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 		}
 	}
 
-	private void checkEnhancement(
+	protected void checkEnhancement(
 			final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> inputInterpolantAutomaton,
 			final INwaOutgoingLetterAndTransitionProvider<L, IPredicate> interpolantAutomaton)
 			throws AutomataLibraryException, AssertionError, AutomataOperationCanceledException {
@@ -914,107 +554,6 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 				predicateUnifier, conservativeSuccessorCandidateSelection, cannibalize);
 	}
 
-	protected void minimizeAbstractionIfEnabled()
-			throws AutomataOperationCanceledException, AutomataLibraryException, AssertionError {
-		final Minimization minimization = mPref.getMinimization();
-		switch (minimization) {
-		case NONE:
-			// do not apply minimization
-			break;
-		case DFA_HOPCROFT_LISTS:
-		case DFA_HOPCROFT_ARRAYS:
-		case MINIMIZE_SEVPA:
-		case SHRINK_NWA:
-		case NWA_MAX_SAT:
-		case NWA_MAX_SAT2:
-		case RAQ_DIRECT_SIMULATION:
-		case RAQ_DIRECT_SIMULATION_B:
-		case NWA_COMBINATOR_PATTERN:
-		case NWA_COMBINATOR_EVERY_KTH:
-		case NWA_OVERAPPROXIMATION:
-		case NWA_COMBINATOR_MULTI_DEFAULT:
-		case NWA_COMBINATOR_MULTI_SIMULATION:
-			// apply minimization
-			minimizeAbstraction(mStateFactoryForRefinement, mPredicateFactoryResultChecking, minimization);
-			break;
-		default:
-			throw new AssertionError();
-		}
-	}
-
-	/**
-	 * Automata theoretic minimization of the automaton stored in mAbstraction. Expects that mAbstraction does not have
-	 * dead ends.
-	 *
-	 * @param predicateFactoryRefinement
-	 *            PredicateFactory for the construction of the new (minimized) abstraction.
-	 * @param resultCheckPredFac
-	 *            PredicateFactory used for auxiliary automata used for checking correctness of the result (if
-	 *            assertions are enabled).
-	 */
-	protected void minimizeAbstraction(final PredicateFactoryRefinement predicateFactoryRefinement,
-			final PredicateFactoryResultChecking resultCheckPredFac, final Minimization minimization)
-			throws AutomataOperationCanceledException, AutomataLibraryException, AssertionError {
-
-		final Function<IPredicate, Set<IcfgLocation>> lcsProvider =
-				x -> (x instanceof ISLPredicate ? Collections.singleton(((ISLPredicate) x).getProgramPoint())
-						: new HashSet<>(Arrays.asList(((IMLPredicate) x).getProgramPoints())));
-		AutomataMinimization<Set<IcfgLocation>, IPredicate, L> am;
-		try {
-			am = new AutomataMinimization<>(getServices(), (INestedWordAutomaton<L, IPredicate>) mAbstraction,
-					minimization, mComputeHoareAnnotation, mIteration, predicateFactoryRefinement,
-					MINIMIZE_EVERY_KTH_ITERATION, mStoredRawInterpolantAutomata, mInterpolAutomaton,
-					MINIMIZATION_TIMEOUT, resultCheckPredFac, lcsProvider, true);
-		} catch (final AutomataMinimizationTimeout e) {
-			mCegarLoopBenchmark.addAutomataMinimizationData(e.getStatistics());
-			throw e.getAutomataOperationCanceledException();
-		}
-		mCegarLoopBenchmark.addAutomataMinimizationData(am.getStatistics());
-		final boolean newAutomatonWasBuilt = am.newAutomatonWasBuilt();
-
-		if (newAutomatonWasBuilt) {
-			// postprocessing after minimization
-			final IDoubleDeckerAutomaton<L, IPredicate> newAbstraction = am.getMinimizedAutomaton();
-
-			// extract Hoare annotation
-			if (mComputeHoareAnnotation) {
-				final Map<IPredicate, IPredicate> oldState2newState = am.getOldState2newStateMapping();
-				if (oldState2newState == null) {
-					throw new AssertionError("Hoare annotation and " + minimization + " incompatible");
-				}
-				mHaf.updateOnMinimization(oldState2newState, newAbstraction);
-			}
-
-			// statistics
-			final int oldSize = mAbstraction.size();
-			final int newSize = newAbstraction.size();
-			assert oldSize == 0 || oldSize >= newSize : "Minimization increased state space";
-
-			// use result
-			mAbstraction = (A) newAbstraction;
-		}
-	}
-
-	@Override
-	protected void computeIcfgHoareAnnotation() {
-		if (mCsToolkit.getManagedScript().isLocked()) {
-			throw new AssertionError("SMTManager must not be locked at the beginning of Hoare annotation computation");
-		}
-		final INestedWordAutomaton<L, IPredicate> abstraction = (INestedWordAutomaton<L, IPredicate>) mAbstraction;
-		mCegarLoopBenchmark.start(CegarLoopStatisticsDefinitions.HoareAnnotationTime.toString());
-		try {
-			new HoareAnnotationExtractor<>(getServices(), abstraction, mHaf);
-			final HoareAnnotationComposer clha = new HoareAnnotationComposer(mCsToolkit, mPredicateFactory, mHaf,
-					getServices(), mSimplificationTechnique, mXnfConversionTechnique);
-			final HoareAnnotationWriter writer = new HoareAnnotationWriter(mIcfg, mCsToolkit, mPredicateFactory, clha,
-					getServices(), mSimplificationTechnique, mXnfConversionTechnique);
-			writer.addHoareAnnotationToCFG();
-			mCegarLoopBenchmark.addHoareAnnotationData(clha.getHoareAnnotationStatisticsGenerator());
-		} finally {
-			mCegarLoopBenchmark.stop(CegarLoopStatisticsDefinitions.HoareAnnotationTime.toString());
-		}
-	}
-
 	@Override
 	public IElement getArtifact() {
 		if (mPref.artifact() == Artifact.ABSTRACTION || mPref.artifact() == Artifact.INTERPOLANT_AUTOMATON
@@ -1046,22 +585,6 @@ public class BasicCegarLoop<L extends IIcfgTransition<?>, A extends IAutomaton<L
 			throw e;
 		} catch (final AutomataLibraryException e) {
 			throw new AssertionError(e);
-		}
-	}
-
-	private static final boolean checkStoreCounterExamples(final TAPreferences pref) {
-		return pref.getMinimization() == Minimization.NWA_OVERAPPROXIMATION;
-	}
-
-	private static SearchStrategy getSearchStrategy(final IPreferenceProvider mPrefs) {
-		switch (mPrefs.getEnum(TraceAbstractionPreferenceInitializer.LABEL_COUNTEREXAMPLE_SEARCH_STRATEGY,
-				CounterexampleSearchStrategy.class)) {
-		case BFS:
-			return SearchStrategy.BFS;
-		case DFS:
-			return SearchStrategy.DFS;
-		default:
-			throw new IllegalArgumentException();
 		}
 	}
