+++ trunk/source/TraceAbstraction/src/de/uni_freiburg/informatik/ultimate/plugins/generator/traceabstraction/CegarLoopSWBnonRecursive.java
@@ -40,10 +40,8 @@ import de.uni_freiburg.informatik.ultimate.automata.AutomataLibraryException;
 import de.uni_freiburg.informatik.ultimate.automata.AutomataLibraryServices;
 import de.uni_freiburg.informatik.ultimate.automata.AutomataOperationCanceledException;
 import de.uni_freiburg.informatik.ultimate.automata.AutomatonEpimorphism;
-import de.uni_freiburg.informatik.ultimate.automata.IAutomaton;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.IDoubleDeckerAutomaton;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.INestedWordAutomaton;
-import de.uni_freiburg.informatik.ultimate.automata.nestedword.INwaOutgoingLetterAndTransitionProvider;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.NestedRun;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.NestedWord;
 import de.uni_freiburg.informatik.ultimate.automata.nestedword.NestedWordAutomaton;
@@ -77,8 +75,7 @@ import de.uni_freiburg.informatik.ultimate.plugins.generator.traceabstraction.pr
  * @author haettigj@informatik.uni-freiburg.de
  *
  */
-public class CegarLoopSWBnonRecursive<L extends IIcfgTransition<?>>
-		extends BasicCegarLoop<L, IAutomaton<L, IPredicate>> {
+public class CegarLoopSWBnonRecursive<L extends IIcfgTransition<?>> extends NwaCegarLoop<L> {
 	/**
 	 * Maps states from the original automaton to corresponding states in the new interpolant automaton.
 	 */
@@ -172,9 +169,10 @@ public class CegarLoopSWBnonRecursive<L extends IIcfgTransition<?>>
 	 * @param services
 	 * @param transitionClazz
 	 */
-	public CegarLoopSWBnonRecursive(final DebugIdentifier name, final IAutomaton<L, IPredicate> initialAbstraction,
-			final IIcfg<?> icfg, final CfgSmtToolkit csToolkit, final PredicateFactory predicateFactory,
-			final TAPreferences taPrefs, final Set<IcfgLocation> errorLocs, final InterpolationTechnique interpolation,
+	public CegarLoopSWBnonRecursive(final DebugIdentifier name,
+			final INestedWordAutomaton<L, IPredicate> initialAbstraction, final IIcfg<?> icfg,
+			final CfgSmtToolkit csToolkit, final PredicateFactory predicateFactory, final TAPreferences taPrefs,
+			final Set<IcfgLocation> errorLocs, final InterpolationTechnique interpolation,
 			final boolean computeHoareAnnotation, final Set<IcfgLocation> hoareAnnotationLocs,
 			final IUltimateServiceProvider services, final Class<L> transitionClazz,
 			final PredicateFactoryRefinement stateFactoryForRefinement) {
@@ -198,10 +196,10 @@ public class CegarLoopSWBnonRecursive<L extends IIcfgTransition<?>>
 
 		// cast the abstraction automaton as nested word and double decker
 		// automaton
-		mNestedAbstraction = (INestedWordAutomaton<L, IPredicate>) mAbstraction;
+		mNestedAbstraction = mAbstraction;
 
-		mDoubleDeckerAbstraction = new RemoveUnreachable<>(new AutomataLibraryServices(getServices()),
-				(INwaOutgoingLetterAndTransitionProvider<L, IPredicate>) mAbstraction).getResult();
+		mDoubleDeckerAbstraction =
+				new RemoveUnreachable<>(new AutomataLibraryServices(getServices()), mAbstraction).getResult();
 		// (IDoubleDeckerAutomaton<LETTER, IPredicate>) mAbstraction.get;
 
 		// cast the path as nested run
